refactor(bridge): extract pure helpers into cohesive sibling modules
bridge.py had grown to 2233 lines, mixing pure helpers with the stateful
Telethon handler wiring. Extract the helper clusters that touch no live
module state (_client/_agent/_my_*/_group_router) into five siblings:

- bridge_topics.py   — forum-topic routing + chat/thread-id parsing
- bridge_approval.py — approval callback encoding + keyboard markup
- bridge_media.py    — voice/image/document detection, download, compose
- bridge_context.py  — transient shared-context formatting
- bridge_commands.py — /persona /stats /aso, model-identity, /osint predicate

bridge.py re-imports every moved name and lists them in __all__, so
kronos.bridge.<name> keeps resolving for callers and tests — the split is
transparent. Constants move with their users: APPROVAL_CALLBACK_PREFIX is
re-exported (the handler registration still needs it); GROQ_* and
_DOC_EXTENSIONS have no outside consumers. Stateful helpers (rate limit,
typing, progress reporter, webhook, handlers) stay put.

Pure move, no behavior change. bridge.py 2233 -> 1624 lines.

Verified: 778 tests pass, ruff clean, import chain + kaos CLI smokes OK.
